♻️ JE-116-refactor: NoReview 컴포넌트 분리
